requirements.txt
boto3==1.28.74
botocore==1.31.74
fabric==3.2.2
invoke==2.2.0
urllib3==2.0.7
flask
random
string
requests
ping3
mysql.connector
mysql-connector-python
